on FABmyTBA Activity, update model prefs in b.g.
As discussed in #342, always close the model settings fragment and
execute the server request in the background. It will Toast after the
request is complete with the status (and color the FAB briefly) to
indicate success or failure.

This couldn't be easily done to the ModelSettingsActivity (which is
launched from the myTBA section for an individual model), because it's a
separate activity that is closed when complete, so there's nothing to
call back to. So we might have to deal with an imperfect UI here for a
bit. Open to better ways to do it...

It'll be doable to make the change, but not in this
timeframe.