Fix #384 Notifications Not Being Marked As Read
See Android Issue 41253: The deleteIntent of the Notification doesn't get
called when the user touches the Notification
https://code.google.com/p/android/issues/detail?id=41253

Workaround: Add a Category to each notification content Intent. When such
an Intent starts an Activity, send the "cancel" Intent that the system
should've sent to dismiss the "read" notification messages.

Contrarywise, don't dismiss notifications when opening the Recent
Notifications activity.

(Possible future: Change the UI so push notifications don't display in a
system notification when Recent Notifications is visible, but still trigger
notification sounds, vibration, and Wear notifications.)