Warn when event not in district for points
Add warning when trying to view district points for a non-district. This
used to be there, but removed for new datafeed simplicity.

HORRIBLE HACK WITHIN: The subscriber returns a custom type that extends
ArrayList so we can use the regular ListViewBinder (because not to would
be to duplicate logic) and the new Binder just casts the data to the
custom type and can access if the event is a district or not there.