Pass EventsByWeek selected tab in onRestoreState
It seemed like fragments are being created twice, one from the
FragmentManager's restoration, and another on in onCreate. Therefore,
save the instance state of the sub-fragment so we can pass a tab to
select on new creation. I don't think this is the best solution, but it
works.... Also not sure why AllTeamsFragment didn't have this issue
either, but oh well...

Fixes #476