Rebind ButterKnife after parent activity stop
It seems like that references to views in the binder were going away
after leaving/coming back to the screen, so it's probably the framework
recreating instances.

BaseAPISubscriber.hasBinderBound is sticky, we we want to reset it when
the parent fragment is stopped so we rebind ButterKnife (and potentially
do other things in the future) if/when we come back.

Fixes #565