Fix live event next/last match bugs
If the eventbus event doesn't give us anything, hide the container for
the views. This fixes the bug in the case where you firt load next/last
match while there are still matches to go, but then resume the activity
after the last match has been played. This will cause the last match
card to be "stale", still showing the old last match. Now we'll hide it
in that case to produce the proper output.

Additionally, check if ButterKnife is bound during the eventbus
callbacks, since we may be in a resume state where we haven't done so
yet. This will fix crashes trying to access a null view in the onEvent
method, which many execute before `updateData` does or before the
framework initially runs `bindViews`.

Fixes #585