Cope with a null ViewModel
ViewModelRenderer. renderToViewModel() is declared @Nullable, so cope
with nulls.

If not for the previous change that makes
DistrictPointsUpdatedNotification return a real ViewModel, each of that
kind of notification saved would trigger this case upon opening the
Recent Notifications screen.

I also put a check in RecentNotificationsListBinder to be safe, but
control doesn’t get there in this case. Perhaps it crashes before then.