Don't Crash on Nullable ViewModels (#784)
* Fix crash bug #780

…at least for the case that I know how to reproduce.

If DistrictPointsUpdatedNotification.renderToViewModel() returns null,
viewing the Recent Notifications screen will crash.

The method is declared @Nullable. That’s inoperative (to use Nixon’s
euphemism for “lie”).

SummaryNotification.renderToViewModel() does return null without
causing a crash, but maybe some other code is careful to not add that
null to the view model list.

* Cope with a null ViewModel

ViewModelRenderer. renderToViewModel() is declared @Nullable, so cope
with nulls.

If not for the previous change that makes
DistrictPointsUpdatedNotification return a real ViewModel, each of that
kind of notification saved would trigger this case upon opening the
Recent Notifications screen.

I also put a check in RecentNotificationsListBinder to be safe, but
control doesn’t get there in this case. Perhaps it crashes before then.

Fixes #780