try to fix the travis CI build
* build-tools-25.0.2 -> 27.0.3
* Follow the instructions in https://docs.travis-ci.com/user/languages/android/

For gradle builds, there are still the following deprecation warnings with deadlines and more. I think some of them come from libraries. See PR #861 for an attempt to fix more of that.

```
Configuration 'debugProguardCompile' is obsolete and has been replaced with 'debugProguardImplementation' and 'debugProguardApi'.
Configuration 'androidTestCompile' is obsolete and has been replaced with 'androidTestImplementation' and 'androidTestApi'.
Configuration 'compile' is obsolete and has been replaced with 'implementation' and 'api'.
Configuration 'androidTestApi' is obsolete and has been replaced with 'androidTestImplementation'.
Configuration 'testCompile' is obsolete and has been replaced with 'testImplementation' and 'testApi'.
Configuration 'debugCompile' is obsolete and has been replaced with 'debugImplementation' and 'debugApi'.
Configuration 'releaseCompile' is obsolete and has been replaced with 'releaseImplementation' and 'releaseApi'.
Configuration 'testApi' is obsolete and has been replaced with 'testImplementation'.
Configuration 'debugBlueCompile' is obsolete and has been replaced with 'debugBlueImplementation' and 'debugBlueApi'.

It will be removed at the end of 2018. For more information see: http://d.android.com/r/tools/update-dependency-configurations.html
```

**Note:** Per [Google Play's target API level requirement](https://developer.android.com/distribute/best-practices/develop/target-sdk)

> Google Play will require that new apps target at least Android 8.0 (API level 26) from August 1, 2018, and that app updates target Android 8.0 from November 1, 2018.

So we can push an app update before Nov. 1. After that, we'll have to update to at least target SDK 26. Last I checked, Robolectric wasn't up to that and there's library version hell.