Migrate team list to ViewPager2
Summary:
ViewPager2 has been released! ViewPager2 is a re-write of ViewPager that now leverages RecyclerView to manage its pages.

This means that updating the number of pages and the content in the tabs is much easier, and we can resolve #706 .

This PR moves only the team list to ViewPager2. Future PRs will move other screens.
The initial work involved was twofold: First, pulling in ViewPager2 and doing the work of replacing the existing ViewPager, and second matching the styling of the new tab strip to the old one.

Issues Reference:

Test Plan:
Manually tested on a variety of devices. I also manually overrode our data loading to simulate loading additional/fewer teams to verify that the tabs and their contents updated appropriately. Seems to be working!

Closes #929